(ios) Updating Xcode project with latest recommendations. Fixing compiler errors. And refactoring large View into it's own file.
