+++ phoenix-ios/phoenix-ios/views/transactions/TxHistoryExporter.swift
@@ -428,12 +428,10 @@ struct TxHistoryExporter: View {
 		exportedCount = 0
 		
 		let databaseManager = Biz.business.databaseManager
-		let peerManager = Biz.business.peerManager
 		let fetcher = Biz.business.paymentsManager.fetcher
 		
 		do {
 			let paymentsDb = try await databaseManager.paymentsDb()
-			let peer = try await peerManager.getPeer()
 			
 			let config = CsvWriter.Configuration(
 				includesFiat: includeFiat,
